requirements.txt
discord.py~=2.5.2
aiohttp~=3.9.5
python-dotenv~=1.0.0
requests~=2.31.0
asyncio~=3.4.3